firewire-request: add more register names
Complete the list of registers that are defined in the CSR architecture.

The CLOCK_* registers are not usually implemented and therefore hidden
by default.  The remaining registers from IEEE-1212 are commented out
because they are reserved in IEEE-1394 (large configuration ROMs are
never accessed indirectly, and the bus uses only the 64-bit fixed
address-space model).