crpp: use binary representation to process image of configuration ROM
In Python 2, str object is similar to bytes object in Python 3, thus
it's possible to handle binary data. On the other hand, in Python 3,
str is an immutable sequence to point to unicode code point, thus
it's not suitable to represent binary data.

Unfortunately, sys.stdin returns str object in both cases, thus
it's required to have condition statements for binary processing.

This commit improves binary operation. Furthermore, struct module is
used to process binary with enough care of endianness.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>